--- src/main/java/com/eyabc/springboot/service/NaverSearchService.java
@@ -1,5 +1,0 @@
-package com.eyabc.springboot.service;
-
-public interface NaverSearchService {
-    public <T> T getByQuery(String query);
-}
