+++ src/main/java/com/eyabc/springboot/service/search/NaverMovieService.java
@@ -1,13 +1,13 @@
-package com.eyabc.springboot.service;
+package com.eyabc.springboot.service.search;
 
 import com.eyabc.springboot.adapter.NaverAdapter;
-import com.eyabc.springboot.dto.Naver.NaverMovieDto;
+import com.eyabc.springboot.dto.naver.NaverMovieDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 @Service
 @RequiredArgsConstructor
-public class NaverMovieService implements NaverSearchService {
+public class NaverMovieService implements SearchService {
     private final NaverAdapter naverAdapter;
 
     @Override
