+++ src/main/java/com/eyabc/springboot/service/search/SearchService.java
@@ -1,0 +1,5 @@
+package com.eyabc.springboot.service.search;
+
+public interface SearchService {
+    public <T> T getByQuery(String query);
+}
